Complete dynamic update functionality
Added code in events.js to the update dynamic button so that it takes the
id of the birdspot's update form, comparing it to the update button and shows
the form of the specific birdspot rather than all birdspots. Also in ui.js,
fixed a typo where there was a parentheses instead of a closing html bracket.