Create home button and add show/hide functions to appropriate elements
Created home button in index.html, also restructured home, changepw, and
sign out buttons to be above both auth and birdspot message elements as
the 3 buttons mentioned previously will be combined into a nav bar.
Createdevent listener for home button on click in app.js. Created hide
and show functions for the home button in bird-spot/events.js. Made it
so that bothindex buttons have a button appearing that leads to the
other index view. Added home button hide function to signout in
auth/ui.js along with view all birdspots as that was not caught before.